[chore] Move otlptext to debug/internal, no need in exporter
