+++ exporter/debugexporter/internal/otlptext/known_sync_error_other.go
@@ -3,7 +3,7 @@
 
 //go:build !linux && !darwin && !windows
 
-package otlptext // import "go.opentelemetry.io/collector/exporter/internal/otlptext"
+package otlptext // import "go.opentelemetry.io/collector/exporter/debugexporter/internal/otlptext"
 
 // knownSyncError returns true if the given error is one of the known
 // non-actionable errors returned by Sync on Plan 9.
