+++ exporter/debugexporter/internal/otlptext/testdata/logs/embedded_maps.out
