+++ exporter/debugexporter/internal/otlptext/testdata/logs/one_record.out
