+++ exporter/debugexporter/internal/otlptext/testdata/logs/two_records.out
