+++ exporter/debugexporter/internal/otlptext/testdata/metrics/invalid_metric_type.out
