+++ exporter/debugexporter/internal/otlptext/testdata/metrics/metrics_with_all_types.out
