+++ exporter/debugexporter/internal/otlptext/testdata/metrics/two_metrics.out
