+++ exporter/debugexporter/internal/otlptext/testdata/profiles/two_profiles.out
