+++ exporter/debugexporter/internal/otlptext/testdata/traces/two_spans.out
