+++ exporter/debugexporter/internal/otlptext/traces.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package otlptext // import "go.opentelemetry.io/collector/exporter/internal/otlptext"
+package otlptext // import "go.opentelemetry.io/collector/exporter/debugexporter/internal/otlptext"
 
 import (
 	"go.opentelemetry.io/collector/pdata/ptrace"
